Mini-charts formatted, bin size preset on dataset load, highcharts chart size bug fixed
Mini-charts are now much better formatted.
The bin size adjusts to ~40 bins/mini-chart. In future versions, we might want to have this default value adjustable in the settings. Note that it's not the exact value, since all record lengths are different.
Highcharts bug - remember when it didn't scale properly to the whole width of the table sell - well, I found a workaround by issuing a 'redraw' event after the mini-charts finish loading.